refactor: utils in separate files, readme update
