+++ .eslintrc.yml
@@ -17,6 +17,7 @@ plugins:
 rules: {
   no-param-reassign: off,
   import/prefer-default-export: off,
+  arrow-body-style: off
 }
 ignorePatterns:
   - /node_modules
