+++ docs/package.json
@@ -13,6 +13,6 @@
     "vite": "^5.0.0"
   },
   "dependencies": {
-    "@fluffio/core": "^0.1.2"
+    "@fluffio/core": "^0.1.3"
   }
 }
