+++ package.json
@@ -35,7 +35,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/geophyzik/sunrise.git"
+    "url": "https://github.com/geophyzik/fluffio.git"
   },
   "keywords": [
     "popover",
