+++ src/utils/get-scroll-size.ts
@@ -1,0 +1,6 @@
+export const getScrollSize = () => {
+  const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
+  const scrollLeft = window.pageXOffset || document.documentElement.scrollLeft;
+
+  return { scrollTop, scrollLeft };
+};
