+++ src/utils/supports-popover.ts
@@ -1,0 +1,3 @@
+export const supportsPopover = () => {
+  return Object.prototype.hasOwnProperty.call(HTMLElement.prototype, 'popover');
+};
