+++ tsconfig.json
@@ -13,7 +13,9 @@
     "noEmit": true,
     "lib": ["es2019", "dom"],
   },
-  "files": ["src/index.ts"],
+  "include": [
+    "src/**/*"
+  ],
   "exclude": [
     "docs"
   ]
